Manage user table replaces the user cards
