--- frontend/components/UserBlock/_styles.scss
@@ -1,180 +1,0 @@
-.user-block {
-  transition: border 300ms ease-in-out, background-color 300ms ease-in-out;
-  background-color: $white;
-  display: inline-block;
-  height: 440px;
-  margin-left: 30px;
-  margin-top: 30px;
-  position: relative;
-  width: 240px;
-  box-shadow: 0 2px 12px 0 rgba(102, 105, 111, 0.12);
-  border: solid 0.5px rgba(156, 163, 172, 0.35);
-  border-radius: 2px;
-
-  &--invited {
-    border: 1px dashed $core-blue;
-  }
-
-  &--disabled {
-    background-color: $core-light-blue-grey;
-  }
-
-  &__header {
-    transition: color 300ms ease-in-out, background-color 300ms ease-in-out;
-    background-color: $core-light-blue-grey;
-    border-bottom: 1px solid $ui-borders;
-    color: $core-black;
-    height: 51px;
-    margin-bottom: $pad-xsmall;
-    text-align: center;
-    width: 100%;
-    border-radius: 2px 2px 0 0;
-
-    &--admin {
-      background-color: $core-blue;
-      color: $white;
-    }
-
-    &--user {
-      background-color: $white;
-      color: $core-black;
-    }
-
-    &--invited {
-      background-color: #f9f0ff;
-      color: #515762;
-    }
-
-    &--disabled {
-      background-color: $core-medium-blue-grey;
-      color: $core-black;
-    }
-  }
-
-  &__header-name {
-    line-height: 51px;
-    margin: 0;
-    padding: 0;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    font-weight: $regular;
-    font-size: $medium;
-  }
-
-  &__details {
-    padding: 0 $pad-large;
-
-    .Select {
-      position: absolute;
-      bottom: $pad-large;
-      width: calc(239px - (#{$pad-large} * 2));
-    }
-  }
-
-  &__avatar {
-    @include size(100px);
-    transition: border 300ms ease-in-out;
-    border: 1px solid $core-dark-blue-grey;
-    filter: none;
-    display: block;
-    margin: 0 auto;
-
-    &--enabled {
-      border: 1px solid $core-blue;
-    }
-  }
-
-  &__status-wrapper {
-    border-top: 1px solid $ui-borders;
-    border-bottom: 1px solid $ui-borders;
-    margin-top: $pad-xsmall;
-    padding: $pad-xsmall 0;
-  }
-
-  &__status-label {
-    transition: color 300ms ease-in-out;
-    float: left;
-    font-weight: $bold;
-    font-size: $medium;
-    letter-spacing: 1px;
-    text-transform: uppercase;
-  }
-
-  &__status-text {
-    transition: color 300ms ease-in-out;
-    float: right;
-    text-transform: uppercase;
-    font-weight: $regular;
-    font-size: $medium;
-    letter-spacing: 1px;
-
-    &--invited {
-      color: $core-blue;
-    }
-
-    &--enabled {
-      color: $success;
-    }
-
-    &--disabled {
-      color: $core-dark-blue-grey;
-    }
-  }
-
-  &__username {
-    transition: color 300ms ease-in-out;
-    color: $core-dark-blue-grey;
-    font-size: $medium;
-    text-transform: uppercase;
-    line-height: $x-large;
-    margin: 0;
-    letter-spacing: 1px;
-
-    &--enabled {
-      color: $core-blue;
-    }
-
-    &--hidden {
-      display: none;
-    }
-  }
-
-  &__position {
-    @include ellipsis;
-    font-size: $small;
-    line-height: $x-large;
-    margin: 0;
-    letter-spacing: 0.5px;
-    color: $core-black;
-    font-weight: $regular;
-
-    &--hidden {
-      display: none;
-    }
-  }
-
-  &__email {
-    transition: color 300ms ease-in-out;
-    color: $core-purple;
-    font-size: $small;
-    line-height: $x-large;
-    margin: 0;
-
-    &--disabled {
-      color: $core-dark-blue-grey;
-    }
-  }
-
-  &__last-seen {
-    font-size: $x-small;
-    line-height: $x-large;
-    margin: 0;
-    color: $core-dark-blue-grey;
-  }
-}
-
-.revoke-invite {
-  position: absolute;
-  width: 189px;
-  bottom: 43px;
-}
