+++ frontend/components/UserRow/helpers.tests.js
