+++ frontend/components/UserRow/index.js
@@ -1,0 +1,1 @@
+export default from './UserRow';
