+++ frontend/pages/admin/UserManagementPage/UserManagementPage.jsx
@@ -17,7 +17,7 @@ import { renderFlash } from 'redux/nodes/notifications/actions';
 import WarningBanner from 'components/WarningBanner';
 import { updateUser } from 'redux/nodes/auth/actions';
 import userActions from 'redux/nodes/entities/users/actions';
-import UserBlock from 'components/UserBlock';
+import UserRow from 'components/UserRow';
 import userInterface from 'interfaces/user';
 
 const baseClass = 'user-management';
@@ -190,7 +190,7 @@ export class UserManagementPage extends Component {
     return false;
   }
 
-  renderUserBlock = (user, idx, options = { invite: false }) => {
+  renderUserRow = (user, idx, options = { invite: false }) => {
     const { currentUser, userErrors } = this.props;
     const { invite } = options;
     const { onEditUser, onToggleEditUser, onUserActionSelect } = this;
@@ -198,7 +198,7 @@ export class UserManagementPage extends Component {
     const isEditing = includes(usersEditing, user.id);
 
     return (
-      <UserBlock
+      <UserRow
         isEditing={isEditing}
         isInvite={invite}
         isCurrentUser={currentUser.id === user.id}
@@ -250,7 +250,7 @@ export class UserManagementPage extends Component {
     return (
       <div className={`${baseClass}__smtp-warning-wrapper`}>
         <WarningBanner
-          message="Email is not currently configured in Fleet. User management features require email."
+          message="SMTP is not currently configured in Fleet. The &quot;Add new user&quot; features requires that SMTP is configured in order to send invitation emails."
           onResolve={goToAppConfigPage}
           shouldShowWarning={!config.configured}
         />
@@ -258,26 +258,40 @@ export class UserManagementPage extends Component {
     );
   }
 
-  renderUsersAndInvites = () => {
+  renderUserTable = () => {
     const { invites, users } = this.props;
-    const { renderUserBlock } = this;
+    const { renderUserRow } = this;
 
     return (
-      <div className={`${baseClass}__users`}>
-        {users.map((user, idx) => {
-          return renderUserBlock(user, idx);
-        })}
-        {invites.map((user, idx) => {
-          return renderUserBlock(user, idx, { invite: true });
-        })}
+      <div className={`${baseClass}__wrapper`}>
+        <table className={`${baseClass}__table`}>
+          <thead>
+            <tr>
+              <th>Username</th>
+              <th>Status</th>
+              <th>Full Name</th>
+              <th>Email</th>
+              <th>Role</th>
+              <th>Position</th>
+              <th />
+            </tr>
+          </thead>
+          <tbody>
+            {users.map((user, idx) => {
+              return renderUserRow(user, idx);
+            })}
+            {invites.map((user, idx) => {
+              return renderUserRow(user, idx, { invite: true });
+            })}
+          </tbody>
+        </table>
       </div>
     );
   }
 
   render () {
-    const { renderModal, renderSmtpWarning, renderUsersAndInvites, toggleInviteUserModal } = this;
-    const { config, invites, loadingInvites, loadingUsers, users } = this.props;
-    const resourcesCount = users.length + invites.length;
+    const { renderModal, renderSmtpWarning, renderUserTable, toggleInviteUserModal } = this;
+    const { config, loadingInvites, loadingUsers } = this.props;
 
     if (loadingInvites || loadingUsers) {
       return false;
@@ -286,20 +300,20 @@ export class UserManagementPage extends Component {
     return (
       <div className={`${baseClass} body-wrap`}>
         <div className={`${baseClass}__heading-wrapper`}>
-          <h1 className={`${baseClass}__user-count`}>Listing {resourcesCount} users</h1>
           <div className={`${baseClass}__add-user-wrap`}>
+            <h1 className={`${baseClass}__user-count`}>Manage users</h1>
             <Button
-              className={`${baseClass}__add-user-btn`}
+              className={`${baseClass}__add-user-btn button button--brand`}
               disabled={!config.configured}
               onClick={toggleInviteUserModal}
               title={config.configured ? 'Add User' : 'Email must be configured to add users'}
             >
-              Add User
+              Add new user
             </Button>
           </div>
         </div>
         {renderSmtpWarning()}
-        {renderUsersAndInvites()}
+        {renderUserTable()}
         {renderModal()}
       </div>
     );
