+++ frontend/styles/var/colors.scss
@@ -22,5 +22,5 @@ $core-blue-ultralight: #edd6ff;
 $button-shadow: #734893;
 $success: #27d77e;
 $success-light: #48da90;
-$warning: #ffad00;
+$warning: #f2c94c;
 $warning-light: #ffda8c;
